api/requirements.txt
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
idna==2.8
importlib-metadata==1.3.0
more-itertools==8.0.2
packaging==19.2
pluggy==0.13.1
py==1.8.0
pyparsing==2.4.5
pytest==5.3.2
pytest-mock==1.13.0
python-chess==0.29.0
requests==2.22.0
six==1.13.0
urllib3==1.25.7
wcwidth==0.1.7
zipp==0.6.0
